libs/py/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "halloween-libs"
version = "0.1.0"
description = "Shared Python libraries and integration tests for Halloween project"
authors = [
  { name = "Thomas Ingeman-Nielsen", email = "thomas@i-n.dk" }
]
readme = "README.md"
requires-python = ">=3.12"

# No runtime dependencies yet, add later as needed
dependencies = []

[project.optional-dependencies]
test = [
  "pytest",
  "paho-mqtt",
  "websockets"
]

[tool.setuptools]
# Use src-style layout: our code lives directly under libs/py
packages = { find = { where = ["."], include = ["halloween_common*"] } }

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = [
  "tests"
]
